requirements.txt
fastapi==0.118.2
uvicorn[standard]==0.37.0
python-multipart==0.0.20

transformers==4.57.0
tokenizers==0.22.1

torch==2.8.0
torchaudio==2.8.0

librosa==0.11.0
soundfile==0.13.1
pyannote.audio==4.0.2

numpy==2.3.3
pandas==2.3.3
matplotlib==3.10.7

python-dotenv==1.1.0